+++ package.json
@@ -2,7 +2,7 @@
   "name": "node-thrift-protocol",
   "version": "0.1.17",
   "description": "An implementation of thrift-protocol with node.",
-  "main": "thrift.js",
+  "main": "index.js",
   "scripts": {
     "lint": "eslint .",
     "mocha": "mocha tests",
